Extracted NavHost and Route to navigation package and changed the FAB into circle shape
