+++ app/src/main/java/tech/relaycorp/letro/utility/ExtensionFunctions.kt
@@ -1,7 +1,7 @@
 package tech.relaycorp.letro.utility
 
 import androidx.navigation.NavController
-import tech.relaycorp.letro.Route
+import tech.relaycorp.letro.ui.navigation.Route
 
 fun NavController.navigateWithPoppingAllBackStack(route: Route) {
     navigate(route.name) {
